backend/requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.15,<4.0
djangorestframework-simplejwt>=5.3,<6.0
django-cors-headers>=4.4,<5.0
django-filter>=24.2,<25.0
psycopg[binary]>=3.2,<4.0
qrcode[pil]>=7.4,<8.0
reportlab>=4.2,<5.0
Pillow>=10.0,<12.0
weasyprint>=62.0,<63.0
